Extract DAL from time_tracker_api #50
