+++ .env.template
@@ -1,5 +1,8 @@
-# Package where the app is located
+# API
+## Package where the app is located
 export FLASK_APP=time_tracker_api
 
-# The database connection URI. Check out the README.md for more details
+
+# Common attributes
+## The database connection URI. Check out the README.md for more details
 DATABASE_URI=mssql+pyodbc://<user>:<password>@time-tracker-srv.database.windows.net/<database>?driver\=ODBC Driver 17 for SQL Server
