+++ requirements/sql_db_serverless.txt
@@ -1,0 +1,12 @@
+# requirements/sql_db_serverless.txt
+
+# For SQL database serverless (MS SQL)
+
+
+# SQL Server driver
+pyodbc==4.0.30
+
+# ORM
+SQLAlchemy==1.3.15
+SQLAlchemy-Utils==0.36.3
+flask_sqlalchemy==2.4.1
