+++ requirements/time_tracker_api/prod.txt
@@ -1,4 +1,8 @@
-# requirements/prod.txt
+# requirements/time_tracker_api/prod.txt
+
+# Dependencies
+-r ../sql_db_serverless.txt
+
 
 # For production releases
 
@@ -20,12 +24,6 @@ Faker==4.0.2
 
 #CLI support
 Flask-Script==2.0.6
-
-# SQL database (MS SQL)
-pyodbc==4.0.30
-SQLAlchemy==1.3.15
-SQLAlchemy-Utils==0.36.3
-flask_sqlalchemy==2.4.1
 
 # Handling requests
 requests==2.23.0
