+++ time_tracker_api/activities/activities_model.py
@@ -6,10 +6,10 @@ class ActivityDao(CRUDDao):
 
 
 def create_dao() -> ActivityDao:
-    from time_tracker_api.sql_repository import db
-    from time_tracker_api.sql_repository import SQLCRUDDao
     from sqlalchemy_utils import UUIDType
     import uuid
+    from commons.data_access_layer.azure.sql_repository import db
+    from commons.data_access_layer.azure.sql_repository import SQLCRUDDao
 
     class ActivitySQLModel(db.Model):
         __tablename__ = 'activity'
